Anemometer Library v0.8 Reads input from the given port and pin for anemometer interrupts.
Required methods:
-----------------------

anemometerInit(anemometerPort,anemometerPortPin);

Initalizes the anemometer module library to read from the specified port and pin.

anemometerStart();

Enables interrupts, beginning capture of speed and buffer updates.

anemometerSpeedBufferRefresh();

Refreshes the buffer by adding a new speed reading to it. Uses the number
of pulses and the time lapse between the refresh to calculate this new
speed.

anemometerSpeedBufferGetAverage();

Returns the average speed in meters per second calculated from the current sum of the buffer.

anemometerSpeedConvertKmH();

Converts the given speed in m/s to km/h

anemometerSpeedConvertMiH();

Converts the given speed in m/s to mi/h.

To use the library:

1. anemometerInit(port,pin)

2. anemometerStart();

3. anemometerSpeedBufferRefresh(); [Updates the buffer and adds new entry]

4. anemometerSpeedBufferGetAverage(); [Returns the current buffer average]

5. anemometerSpeedConvert[KmH|MiH](); [Converts the speed from ms to kmh
or mih]