I2C Library v0.3 and BMP Library 0.5 I2C Library 0.3 ----------------- Only requires using the i2cInit(ab_i2c_module_#), once this is done, the i2c is working at a minimum level. Functionality to be extended.
BMP Library 0.5
----------------
Requires the use of the bmpInit(ab_i2c_module_#) to initalize the library
to an i2c module. The number of bmpInit() functions called determines the
number of different BMP Modules to interface, they are stored in an
increasing order until a maximum of 20. Once set, bmpDataRead(index) must
be called to read from the instantiation # module of the library. Index
determines the module to read from. As of version 0.5, the pin regarding
to the BMP module must be turned off manually.

Have a nice day! :D