BinCounter v0.9 and Buttons Library v0.9 BinCounter Library v0.9:
Requires only the use of binCounterInit(clkPort, clkPin, func0Port,
func0Pin, func1Port, func1Pin) to enable the binary counter. Once inited,
only the binCounterIncrease(), binCounterDecrease(), binCounterClear()
functions are used to control the counter.

Buttons Library v0.9:

Requires only the use of buttonsInit(). By default, the buttons are
initialized on Port B as required for Aerobal Project. Once inited, you
can check if there was a button pressed with the buttonsWasPressed()
function and independently with the buttonsWasPressedB[0,1,2,3,4,5]().
The buttons masks for interrupts can be disabled with the
buttonsDisabled() and reenabled with the buttonsEnable().
buttonsInterruptHandler() handles the setting of which button was
pressed.